requirements.txt
#cuda==11.1
librosa==0.9.1
tensorboardX==2.5
tensorboard
pandas
asteroid-filterbanks
fastapi
uvicorn
python-dotenv
requests
pytest
httpx